+++ pyforestscan_qgis/algorithms/advanced/__init__.py
@@ -7,6 +7,7 @@ from .advanced_point_density import AdvancedPointDensityAlgorithm
 from .advanced_pai import AdvancedPaiAlgorithm
 from .advanced_canopy_cover import AdvancedCanopyCoverAlgorithm
 from .advanced_fhd import AdvancedFhdAlgorithm
+from .ept_subset import EptSubsetExtractAlgorithm
 from .advanced_rumple import AdvancedRumpleAlgorithm
 from .advanced_voxel_stat import AdvancedVoxelStatAlgorithm
 from .normalize_hag import NormalizeHagAlgorithm
@@ -20,6 +21,7 @@ __all__ = [
     "AdvancedPaiAlgorithm",
     "AdvancedCanopyCoverAlgorithm",
     "AdvancedFhdAlgorithm",
+    "EptSubsetExtractAlgorithm",
     "AdvancedRumpleAlgorithm",
     "AdvancedVoxelStatAlgorithm",
     "NormalizeHagAlgorithm",